Reset MaxSinceStart XML element value when on ground
The MaxSinceStart XML element is used to compute the "Altitude Loss
After Takeoff" by subtracting the current altitue from the maximum
altitude since start.

This approach triggers a false TERRAIN / DON'T SINK alert when flying
circuits / patterns or when performing touch and go landings.

To fix this issue reset the maximum value of the MaxSinceStart XML
element when we are on ground.

Note: it would probably be better to rename the MaxSinceStart tag to
MaxSinceTakeoff, however that would require changing many planes files
and potentially cause conflicts with subsequent updates from Asobo.