Save ATC options
Save the tail number, call sign and flight number (per aircraft) in
persistent data storage and retrive it when the simulator starts to
avoid setting it every time.

Note that show tail number and append heavy are not yet saved since it is
not possible to distinguish beetween option being true/false or not set.