Stay on last page that has data for each column (left/right)
Do not scroll a column (e.g. stars) further down, if there is no more data to show, even if we scroll down the other column (e.g. approaches or transistions).
Stay on the last page that has data.
Applies to both arrival and departure pages
Small font for all data items (except selected items, which are green and not small)