Fix annunciator buttons on the Longitude
Now that the annunciators are working on the Longitude with SU6, it became really bothersome that the buttons did not work. After some digging I found that the issue is that the logic for managing the annunciations is in the `Cabin_Annunciations` class. The caution/warning reset buttons work on the TBM because its MFD uses that class (in the form of `Engine_Annunciations`). The Longitude however does not have annunciations on the MFD. It needs to use `Cabin_Annunciations` on the PFD instead.